now has wallet for every user. can purchase items with it.

+++ src/main/resources/db/migration/V20250520__wallet_tasks_relationships_tables.sql
@@ -1,7 +1,7 @@
 CREATE TABLE "wallet" (
     id BIGINT AUTO_INCREMENT PRIMARY KEY NOT NULL,
     user_id BIGINT NOT NULL,
-    balance BIGINT DEFAULT 0,
+    balance DOUBLE PRECISION DEFAULT 0,
     FOREIGN KEY (user_id) REFERENCES "user"(id)
 );
